Refactored existing MDX separation into its own class in prep for adding VR arch
