+++ audio_separator/separator/architectures/__init__.py
@@ -1,0 +1,1 @@
+from .mdx_separator import MDXSeparator
